* 'staging' of https://bitbucket.org/devatsrs/neon.web:
  can not edit cc on ticket edit page #1319
  LCR Bug fixed #1086
  NEON-1341 Sippy Account Name - Authentication Rule
  4.17 CDR BIGINT change
  NEON-1341 Sippy Account Name - Authentication Rule

+++ dbv-2/data/revisions/4.17/dev.sql
@@ -50,7 +50,7 @@ BEGIN
 		connect_time datetime,
 		disconnect_time datetime,
 		is_inbound tinyint(1) default 0,
-		ID INT,
+		ID BIGINT(20),
 		ServiceID INT
 	);
 	INSERT INTO tmp_tblUsageDetails_
@@ -141,7 +141,7 @@ BEGIN
 			connect_time datetime,
 			disconnect_time datetime,
 			is_inbound tinyint(1) default 0,
-			ID INT
+			ID BIGINT(20)
 	);
 	INSERT INTO tmp_tblUsageDetails_
 	SELECT
